+++ internal/helixclient/client.go
@@ -3,7 +3,7 @@ package helixclient
 import (
 	"errors"
 
-	"github.com/nicklaw5/helix"
+	"github.com/nicklaw5/helix/v2"
 	"github.com/zneix/tcb2/internal/config"
 )