+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 ## Unversioned
 
-## 2.0
+## 2.0-rc2
+
+- Minor: Reimplemented channel MOTDs. (#10)
+- Minor: Added a way to subscribe to all events at once with `!notifyme all`.
+- Minor: List all available events in `!notifyme` command.
+- Minor: Reimplemented respecting of disabled command.
+- Minor: Make periodic requests to supinic's API to signal bot being alive. (#6)
+- Bugfix: Corrected response in `!bot` command.
+- Bugfix: Make checks for subscription values case-insensitive again.
+- Dev: Use millisecond precision in logs.
+- Dev: Added licese. (#7)
+- Dev: Cleaned up code for API routes.
+
+
+## 2.0-rc1
 
 - Initial release.